+++ src/main/java/cn/staitech/anno/service/PathologicalIndicatorCategoryService.java
@@ -94,7 +94,7 @@ public interface PathologicalIndicatorCategoryService {
     /**
      * 查询病例指标下的标注类别数量
      */
-    Integer selectCategoryNumber(Long indicatorId);
+    Long selectCategoryNumber(Long indicatorId);
 
     /**
      * 根据projectId查询标注类别
